implemented all review comments (reduce duplication2)
