+++ src/views/ModifyCommercialView.vue
@@ -3,6 +3,8 @@ import { ref, computed, onMounted } from 'vue';
 import { useRouter } from 'vue-router';
 import { commercialService, type CommercialUnit } from '@/services/CommercialService';
 import { useToast } from 'primevue/usetoast';
+import { handleCancel, showSavingErrorToast, showValidationErrorToast } from '@/helper/viewHelper';
+
 
 const props = defineProps<{
   projectId: string;
@@ -112,12 +114,7 @@ onMounted(() => {
 
 const save = async () => {
   if (!isValid.value) {
-    toast.add({
-      severity: 'error',
-      summary: 'Validierungsfehler',
-      detail: validationErrors.value.join('\n'),
-      life: 6000,
-    });
+    showValidationErrorToast(toast, validationErrors.value);
     return;
   }
 
@@ -140,28 +137,11 @@ const save = async () => {
     router.push(`/project/${props.projectId}/commercial/${props.unitId}`);
   } catch (err) {
     console.error('Fehler beim Speichern:', err);
-    toast.add({
-      severity: 'error',
-      summary: 'Speicherfehler',
-      detail: 'Gewerbe konnte nicht gespeichert werden.',
-      life: 6000,
-    });
+    showSavingErrorToast(toast, 'Gewerbe konnte nicht gespeichert werden.');
   }
 };
 
-const cancel = () => {
-  if (hasChanges.value) {
-    const confirmLeave = confirm(
-      'Es gibt ungespeicherte Änderungen. Möchten Sie die Seite wirklich verlassen?',
-    );
-    if (!confirmLeave) return;
-  }
-  if (window.opener) {
-    window.close();
-  } else {
-    router.push(`/project/${props.projectId}/objects`);
-  }
-};
+const cancel = () => handleCancel(hasChanges, router, props.projectId);
 </script>
 
 <template>
