+++ src/views/ModifyPropertyView.vue
@@ -3,6 +3,8 @@ import { computed, onMounted, ref } from 'vue';
 import { useRouter } from 'vue-router';
 import { propertyService, type PropertyUnit } from '@/services/PropertyService';
 import { useToast } from 'primevue/usetoast';
+import { handleCancel, showSavingErrorToast, showValidationErrorToast } from '@/helper/viewHelper';
+
 
 const props = defineProps<{
   projectId: string;
@@ -185,12 +187,7 @@ onMounted(() => {
 
 const save = async () => {
   if (!isValid.value) {
-    toast.add({
-      severity: 'error',
-      summary: 'Validierungsfehler',
-      detail: validationErrors.value.join('\n'),
-      life: 6000,
-    });
+    showValidationErrorToast(toast, validationErrors.value);
     return;
   }
 
@@ -216,28 +213,11 @@ const save = async () => {
     router.push(`/project/${props.projectId}/property/${props.unitId}`);
   } catch (err) {
     console.error('Fehler beim Speichern:', err);
-    toast.add({
-      severity: 'error',
-      summary: 'Speicherfehler',
-      detail: 'Eigentum konnte nicht gespeichert werden.',
-      life: 6000,
-    });
+    showSavingErrorToast(toast, 'Eigentum konnte nicht gespeichert werden.');
   }
 };
 
-const cancel = () => {
-  if (hasChanges.value) {
-    const confirmLeave = confirm(
-      'Es gibt ungespeicherte Änderungen. Möchten Sie die Seite wirklich verlassen?',
-    );
-    if (!confirmLeave) return;
-  }
-  if (window.opener) {
-    window.close();
-  } else {
-    router.push(`/project/${props.projectId}/objects`);
-  }
-};
+const cancel = () => handleCancel(hasChanges, router, props.projectId);
 </script>
 
 <template>
